requirements.txt
streamlit==1.29.0
pandas==1.4.2
numpy==1.22.3
matplotlib==3.8.0
scikit-learn==1.3.2
Pillow==10.2.0
python-dateutil==2.8.2
